requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
python-dotenv==1.0.1

SQLAlchemy==2.0.29
pymysql==1.1.0

google-genai==0.2.2

PyPDF2==3.0.1

requests==2.31.0

Werkzeug==3.0.1

# google-adk and google-genai will be added as per your environment
